fix- follower feature
